% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AVS.R
\name{AVS}
\alias{AVS}
\title{Active Variable Selection (Via First Order Linear Model)}
\usage{
AVS(x, fx, scale = 0, k = 2, trace = 0)
}
\arguments{
\item{x}{the set of training points in the input space, given as a dataframe.}

\item{fx}{experiment/model evaluation at the training points, given as a vector.}

\item{scale}{specifies the estimate of the error variance.  The default, \code{0}, indicates that this value is to be selected via maximum likelihood.
\code{scale_maxlrm1} indicates that it is to be estimated as the variance parameter of the full model.}

\item{k}{allows specification of the k parameter of the information criterion.
The default, \code{2}, is AIC.  \code{log(n)} is BIC.  Can also specify strings \code{"BIC"} or \code{"logn/2"}.}

\item{trace}{
    if positive, information is printed during the running of \code{step}.
    Larger values may give more detailed information.
  }
}
\value{
\item{stepwise_selected_model}{the stepwise selected model from \code{step}}
\item{active_variables}{the active variables, given by those selected for \code{stepwise_selected_model}}
}
\description{
A method for selecting the active variables of a model by fitting a first-order linear model
using some stepwise selection criteria, and choosing those that are chosen within the model.
}
\examples{
data(USArrests)
X <- USArrests[,c("Assault", "UrbanPop")]
y <- USArrests[,"Murder"]
AVS( X, y )
}
\seealso{
\code{\link[stats]{step}}
}
