% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CL_ML.R
\name{CL_ML}
\alias{CL_ML}
\title{Maximum likelihood for finding Correlation length parameters}
\usage{
CL_ML(
  CF = GaussianCF,
  CF_para = list(),
  para_to_optim,
  x,
  fx,
  G = NA,
  mean_function_model = NA,
  initial,
  lower = -Inf,
  upper = Inf,
  method = "Nelder-Mead"
)
}
\arguments{
\item{CF}{A function that can be used to calculate the correlation between two matrices or dataframes of points.}

\item{CF_para}{additional parameters to the function \code{CF}, along with their specified values, given as a list.}

\item{para_to_optim}{a vector of the names of the arguments to the function \code{CF} that which to assessed via maximum likelihood.}

\item{x}{a matrix of points (given by the rows)}

\item{fx}{output for each \code{x}.}

\item{G}{model matrix.  This can either be specifed here or the code can calculate this given \code{mean_function_model}.}

\item{mean_function_model}{this is a linear model that is used to calculate a model matrix \code{G}.}

\item{initial}{initial settings for the parameter values to be optimised.  These must either be given as a vector or a list of vectors, depending on how many arguments want to be optimised.}

\item{lower}{lower limits for the optimisation, given as a vector or list of vectors.}

\item{upper}{upper limits for the optimisation, given as a vector or list of vectors.}

\item{method}{method choice for optimisation.}
}
\value{
a list containing the found optimium correlation length parameter values.
}
\description{
Maximum likelihood for finding Correlation length parameters
}
\examples{
f <- function(x){ c( x[2] * sin( x[1] ) + x[1] * cos( x[2] ), 2 * x[1] + 3 * x[2] / x[1] ) }
x <- matrix( runif( 20, 0.2, 1.2 ), ncol = 2 )
fx <- t( apply( x, 1, f ) )
CL_ML( CF = NetworkPPBLE::GaussianCF,
       CF_para = list( delta = 0.0001 ),
       para_to_optim = "theta",
       x = x,
       fx = fx,
       mean_function_model = 1,
       initial = rep( 1, 2 )
        )
}
\seealso{
\code{\link[stats]{optim}}
}
