% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulate.R
\name{emulate}
\alias{emulate}
\title{PPBLE emulate}
\usage{
emulate(
  x,
  fx,
  mean_function_model = 1,
  s2 = NA,
  CF = NetworkPPBLE::GaussianCF,
  CF_para = list(),
  CF_para_optim = NULL,
  CF_para_optim_para = list()
)
}
\arguments{
\item{x}{set of training points in the input space, given as the rows of a matrix or dataframe}

\item{fx}{model output corresponding to the training points in \code{x}, given as a vector or matrix.}

\item{mean_function_model}{model used to define the basis functions of the regression polynomial.  If given as \code{1} (default) the regression matrix G is taken to be (1,x).  If given as \code{2} or \code{3}, the regression matrix G is taken to include the full quadratic or cubic terms (respectively) of the quantities in x.  Otherwise a model can be specified from which the regression matrix G can be calculated.}

\item{s2}{a vector of length equal to the number of columns of \code{fx} giving the scalar variance parameter for each output component.}

\item{CF}{a correlation function to be used by the emulator - it is assumed that the first two arguments of this function are to be objects (vectors, matrices or dataframes) for which the correlation function is to be computed.}

\item{CF_para}{specification of the value of any additional parameters that are required for the correlation function, given in the form of a list.  If a value is specified for a particular parameter, this value is used; if \code{NA} is specified, the function \code{CF_para_optim} will be used to optimise the values of those parameters; any additional parameters to \code{CF} not listed will take their default values as given by \code{CF} itself.  Note that any parameters that are not given default values by \code{CF} must be listed.}

\item{CF_para_optim}{a function that optimises some of the parameters of the function \code{CF}.}

\item{CF_para_optim_para}{a list of any parameters to the function \code{CF_para_optim} that require specifying, along with their desired specified values.  Any parameters of \code{CF_para_optim} not listed will take their default value (as given by \code{CF_para_optim}).  Note that any parameters that are not given default values by \code{CF_para_optim} must be listed.}
}
\value{
an object of class \code{emulate} which is a PPBLE that can be used to predict model output at further inputs, along with providing an uncertainty estimate.
}
\description{
generate an object of class \code{emulate} given a set of training points and the corresponding output of a model at those training points.
}
\examples{
f <- function(x){ c( x[2] * sin( x[1] ) + x[1] * cos( x[2] ), 2 * x[1] + 3 * x[2] / x[1] ) }
x <- matrix( runif( 20, 0.2, 1.2 ), ncol = 2 )
fx <- t( apply( x, 1, f ) )
theta <- c(0.4, 0.6)
emulator <- emulate( x = x, fx = fx, CF = GaussianCF,
                  CF_para = list( theta = theta, delta = 0.0001 ) )
emulator2 <- emulate( x = x, fx = fx[,1], CF_para = list( theta = theta, delta = 0.0001 ) )
X <- data.frame( x )
dimnames( X )[[2]] <- c( "X1", "X2" )
emulator3 <- emulate( x = X, fx = fx, CF = GaussianCF,
                   CF_para = list( theta = theta, delta = 0.0001 ) )
emulator4 <- emulate( x = x, fx = fx, CF = GaussianCF,
                      CF_para = list( theta = NA, delta = 0.0001 ),
                      CF_para_optim = CL_ML,
                      CF_para_optim_para = list( mean_function_model = 1,
                                                 initial = rep( 1, 2 ) ) )
}
