% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESplot.R
\name{ESplot}
\alias{ESplot}
\title{ESplot}
\usage{
ESplot(
  fx,
  Efx,
  Varfx,
  n = 3,
  col_Var = "blue",
  col_E = "red",
  col_line = "black",
  main = NULL,
  sub = NULL,
  xlab = "x",
  ylab = "",
  cex = 1,
  cex.axis = 1,
  cex.lab = 1,
  cex.main = 1,
  cex.sub = 1,
  xlim = NULL,
  ylim = NULL,
  pch = 16,
  code = 3,
  angle = 90,
  length = 0.05,
  lwd = 1,
  export_type = NULL,
  filename = NA,
  height = 1000,
  width = 1000,
  mar = (c(5, 4, 4, 2) + 0.1),
  mgp = c(3, 1, 0)
)
}
\arguments{
\item{fx}{simulator evaluations / data, given as a vector}

\item{Efx}{emulator expectations, given as a vector}

\item{Varfx}{emulator variances, given as a vector}

\item{n}{number of standard deviations either side of the mean for the range of the vertical arrow bars.}

\item{col_Var}{colour of the error bars}

\item{col_E}{colour of the points used to represent emulator expectation}

\item{col_line}{colour of the line \code{fx} = \code{Efx}}

\item{main}{main title of the plot}

\item{sub}{subtitle of the plot}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{cex}{size of the points, as given by \code{par}.}

\item{cex.axis}{the magnification to be used for axis annotation relative to the current setting of \code{cex}.}

\item{cex.lab}{the magnification to be used for the x and y labels relative to the current setting of \code{cex}.}

\item{cex.main}{The magnification to be used for the main titles relative to the current setting of \code{cex}.}

\item{cex.sub}{The magnification to be used for sub-titles relative to the current setting of \code{cex}/}

\item{xlim}{x-axis limits}

\item{ylim}{y-axis limits}

\item{pch}{integer specifying a symbol or a single character to be used as the default in plotting points.}

\item{code}{integer code, determining \emph{kind} of arrows to be
    drawn.}

\item{angle}{angle from the shaft of the arrow to the edge of the
    arrow head.}

\item{length}{length of the edges of the arrow head (in inches).}

\item{lwd}{\link[graphics]{graphical parameters}, possible vectors.
    \code{NA} values in \code{col} cause the arrow to be omitted.}

\item{export_type}{type of file that the image should be exported as (if required).  By default this is \code{NULL},
in which case the required plot is plotted in the R interface.}

\item{filename}{name of the exported file.}

\item{height}{height of the image}

\item{width}{width of the image}

\item{mar}{a numerical vector of the form \code{c(bottom, left, top, right)} which gives the number of lines of margin to be specified on the four sides of the plot.  The default is \code{c(5, 4, 4, 2) + 0.1}.}

\item{mgp}{The margin line (in \code{mex} units) for the axis title, axis labels and axis line. Note that \code{mgp[1]} affects title whereas \code{mgp[2:3]} affect \code{axis}. The default is c(3, 1, 0).}
}
\value{
A plot, either exported or in the R interface itself.
}
\description{
ESplot
}
\examples{
fx <- stats::runif( 10 )
Efx <- stats::runif( 10 )
Varfx <- abs( stats::rnorm( 10, sd = 0.05 ) )
ESplot( fx = fx, Efx = Efx, Varfx = Varfx )
}
