% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLSbeta.R
\name{GLSbeta}
\alias{GLSbeta}
\title{Generalised Least Squares Estimate for Beta}
\usage{
GLSbeta(fx, C, G)
}
\arguments{
\item{fx}{a set of n model runs, each with a k-component model output, given as an n by k matrix or n-vector.}

\item{C}{correlation matrix between the model runs \code{fx}.}

\item{G}{model matrix for the set of runs \code{fx}.}
}
\value{
\item{betahatGLS}{Generalised Least Squares estimate for beta coefficients}
\item{Q}{Q matrix - corresponding to G^TC^{-1}G}
}
\description{
Generalised Least Squares Estimate for Beta
}
\examples{
f <- function( x ){ x[2] * sin( x[1] ) + x[1] * cos( x[2] ) }
x <- matrix( runif( 20, 0.2, 1.2 ), ncol = 2 )
fx <- apply( x, 1, f )
C <- GaussianCF( x, theta = c(0.4, 0.6) )
G <- cbind( rep( 1, 10 ), x )
GLSbeta( fx, C, G )
}
