% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaternCF.r
\name{MaternCF}
\alias{MaternCF}
\title{Matern Correlation Function}
\usage{
MaternCF(X, Y = X, kappa, nu)
}
\arguments{
\item{X}{a vector, matrix or dataframe}

\item{Y}{a vector, matrix or dataframe}

\item{kappa}{range paraneter}

\item{nu}{shape parameter}
}
\value{
Matern correlation function value between the rows of X and Y,
given as a matrix of dimension \code{nrow(X)} by \code{nrow(Y)}.
}
\description{
Calculate the Gaussian correlation function between the points in (given by the rows of) two matrices.
}
\examples{
X <- matrix( rnorm( 10 ), ncol = 2 )
Y <- matrix( runif( 6 ), ncol = 2 )
kappa <- 10
nu <- 1/5
MaternCF( X, Y, kappa, nu )
MaternCF( as.data.frame(X), Y, kappa, nu )
}
