% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_sd_plot.r
\name{mean_sd_plot}
\alias{mean_sd_plot}
\title{Mean and standard deviation plot (against continuous x-variable).}
\usage{
mean_sd_plot(
  x,
  fx = NULL,
  Efx,
  Varfx,
  col_Efx = 1,
  col_fx = 2,
  col_sd = 3,
  lwd = 2,
  lty_Efx = 1,
  lty_fx = 1,
  lty_sd = 1,
  ylim = c(0, max(Efx)),
  xlim = c(min(x), max(x)),
  xlab = "",
  ylab = "",
  ca = 1,
  cl = 1,
  main = rep("", nrow(Efx)),
  export_type = NULL,
  filename = NA,
  height = 1000,
  width = 1000,
  mar = (c(5, 4, 4, 2) + 0.1),
  mgp = c(3, 1, 0),
  mfrow = c(1, 1)
)
}
\arguments{
\item{x}{Vector of x-values (typically representing say, time).}

\item{fx}{Matrix of true model output across \code{x}, if known (by default \code{NULL}.}

\item{Efx}{Matrix of emulator expectation across \code{x}.}

\item{Varfx}{Matrix of emulator variance across \code{x}.}

\item{col_Efx}{Colour of mean line.}

\item{col_fx}{Colour of true model line.}

\item{col_sd}{Colour of standard deviation line.}

\item{lwd}{line widths.}

\item{lty_Efx}{Line type for mean line.}

\item{lty_fx}{Line type of true model line.}

\item{lty_sd}{Line type of standard deviation line.}

\item{ylim}{Limits on \code{y}-axis - can be given as an \code{n} by \code{2} matrix if different limits are required for each plot.}

\item{xlim}{Limits on \code{x}-axis - can be given as an \code{n} by \code{2} matrix if different limits are required for each plot.}

\item{xlab}{\code{x}-axis label.}

\item{ylab}{\code{y}-axis label.}

\item{ca}{Font size of axis.}

\item{cl}{Font size of labels.}

\item{main}{vector of plot titles if required.}

\item{export_type}{type of file that the image should be exported as (if required).  By default this is \code{NULL},
in which case the required plot is plotted in the R interface.}

\item{filename}{name of the exported file.}

\item{height}{height of the image}

\item{width}{width of the image}

\item{mar}{a numerical vector of the form \code{c(bottom, left, top, right)} which gives the number of lines of margin to be specified on the four sides of the plot.  The default is \code{c(5, 4, 4, 2) + 0.1}.}

\item{mgp}{The margin line (in \code{mex} units) for the axis title, axis labels and axis line. Note that \code{mgp[1]} affects title whereas \code{mgp[2:3]} affect \code{axis}. The default is c(3, 1, 0).}

\item{mfrow}{The number of rows and columns to divide the device window into, given as a vector of two elements.}
}
\value{
Plot.
}
\description{
Plot the mean and standard deviation curves over a continuous x-variable (typically time).
}
\examples{
x = 1:100
Efx = matrix( c( x^(2.1/2), x^(2.2/2) ), nrow = 2, byrow=TRUE )
Varfx = matrix( c(x,x), nrow = 2, byrow = TRUE)
mean_sd_plot( x = x, Efx = Efx, Varfx = Varfx )
}
