% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_matrix.r
\name{model_matrix}
\alias{model_matrix}
\title{Construct model matrix for any dataframe and model.}
\usage{
model_matrix(x, model = 1)
}
\arguments{
\item{x}{set of training points in the input space, given as the rows of a matrix or dataframe.}

\item{model}{model used to define the basis functions of the regression polynomial.  If given as \code{1} (default) the regression matrix G is taken to be (1,x).  If given as \code{2} or \code{3}, the model matrix \code{MM} is taken to include the full quadratic or cubic terms (respectively) of the quantities in x.  Otherwise a model can be specified from which the regression matrix G can be calculated.}
}
\value{
Model matrix.
}
\description{
Construct model matrix for any dataframe and model.
}
\examples{
x <- matrix( runif( 20, 0.2, 1.2 ), ncol = 2 )
model_matrix( x )
model_matrix( x, 2 )

}
