% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_test.r
\name{partition.test}
\alias{partition.test}
\title{Test Data Partitions}
\usage{
partition.test(x, tgp_object)
}
\arguments{
\item{x}{Matrix of input points to partition.}

\item{tgp_object}{An object of the class \code{tgp}.}
}
\value{
Submatrices of \code{x}, partitioned according to the partitions of \code{tgp_object}.
}
\description{
Find the partitions of a set of test data as given by a Treed Gaussian process having used, for example, \code{\link[tgp]{btgp}}
}
\examples{
# construct some 1-d nonstationary data - this exammple uses one of
# those found in the tgp package itself.
X <- seq( 0, 20, length = 100 )
XX <- seq( 0, 20, length = 99 )
Z <- ( sin( pi * X / 5 ) + 0.2 * cos( 4 * pi * X / 5 ) ) * ( X <= 9.6 )
lin <- X > 9.6;
Z[lin] <- -1 + X[lin] / 10
Z <- Z + rnorm( length( Z ), sd = 0.1 )
out <- tgp::btgp( X = X, Z = Z, XX = XX ) 	# use a treed GP
plot( out ) 			# plot the surface
tgp::tgp.trees( out )
partition.test( x = matrix( XX, ncol = 1 ), tgp_object = out )
}
\seealso{
\code{\link[tgp]{btgp}}
}
