% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeriodicGaussianCF.R
\name{PeriodicGaussianCF}
\alias{PeriodicGaussianCF}
\title{Periodic Gaussian Correlation Function}
\usage{
PeriodicGaussianCF(X, Y = X, theta, p = rep(NA, NCOL(X)), delta = 0)
}
\arguments{
\item{X}{a vector, matrix or dataframe}

\item{Y}{a vector, matrix or dataframe}

\item{theta}{a vector of correlation length parameter values (one for each column of X).}

\item{p}{vector of period values for each of the inputs.  If there is no period then the corresponding element
of this vector should be given as \code{NA}.}

\item{delta}{an (optional) scalar nugget parameter.}
}
\value{
Periodic Gaussian correlation function value between the rows of X and Y,
given as a matrix of dimension \code{nrow(X)} by \code{nrow(Y)}.
}
\description{
Calculate the Gaussian correlation function between the points in (given by the rows of) two matrices, given that
there is periodicity in the values of (some of) the inputs.
}
\examples{
X <- matrix( rnorm( 10 ), ncol = 2 )
Y <- matrix( runif( 6 ), ncol = 2 )
theta <- c( 0.5, 0.8 )
GaussianCF( X, Y, theta = theta )
# if p is not specified, then the results should be the same as above.
PeriodicGaussianCF( X, Y, theta = theta )
# compare if we have a period of 1 for each input.
PeriodicGaussianCF( X, Y, theta = theta, p = rep( 1, 2 ) )
# or a period for just one of the inputs.
PeriodicGaussianCF( X, Y, theta = theta, p = c( NA, 0.5 ) )
}
\seealso{
\code{\link{GaussianCF}}
}
