% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMarrows.R
\name{PMarrows}
\alias{PMarrows}
\title{Plot plus-and-minus standard deviation Arrows}
\usage{
PMarrows(
  x,
  mean,
  Var,
  n = 3,
  code = 3,
  angle = 90,
  length = 0.05,
  col = "black",
  lwd = 1
)
}
\arguments{
\item{x}{vector of x coordinates}

\item{mean}{vector of expected values}

\item{Var}{vector of variances}

\item{n}{number of standard deviations for the vertical arrows to be covering.}

\item{code}{integer code, determining \emph{kind} of arrows to be
    drawn.}

\item{angle}{angle from the shaft of the arrow to the edge of the
    arrow head.}

\item{length}{length of the edges of the arrow head (in inches).}

\item{col}{\link[graphics]{graphical parameters}, possible vectors.
    \code{NA} values in \code{col} cause the arrow to be omitted.}

\item{lwd}{\link[graphics]{graphical parameters}, possible vectors.
    \code{NA} values in \code{col} cause the arrow to be omitted.}
}
\value{
plots some arrows
}
\description{
Plot plus-and-minus standard deviation Arrows
}
\examples{
x <- stats::runif( 10 )
mean <- stats::runif( 10 )
Var <- abs( stats::rnorm( 10, sd = 0.05 ) )
plot( x = x, y = mean, type="n" )
PMarrows( x = x, mean = mean, Var = Var )
}
