% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PowerCF.R
\name{PowerCF}
\alias{PowerCF}
\title{Power Correlation Function}
\usage{
PowerCF(X, Y = X, theta, P = 2, delta = 0)
}
\arguments{
\item{X}{vector, matrix or dataframe}

\item{Y}{a vector, matrix or dataframe}

\item{theta}{a vector of correlation length parameter values (one for each column of X).}

\item{P}{a scalar power parameter, note the default value of 2 yields the Gaussian Correlation Function.}

\item{delta}{an (optional) scalar nugget parameter.}
}
\value{
Power correlation function value between the rows of \code{X} and \code{Y},
given as a matrix of dimension \code{nrow(X)} by \code{nrow(Y)}.
}
\description{
Power Correlation Function
}
\examples{
X <- matrix( rnorm( 10 ), ncol = 2 )
Y <- matrix( runif( 6 ), ncol = 2 )
theta <- c( 0.5, 0.8 )
P <- 1.9
PowerCF( X, Y, theta, P, delta = 0.001 )
}
