% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.emulate}
\alias{predict.emulate}
\title{PPBLE predict (emulate)}
\usage{
\method{predict}{emulate}(object, x = "training", batch_size = 100, ...)
}
\arguments{
\item{object}{an object of class \code{emulate}}

\item{x}{a set of training runs at which to make predictions using the \code{emulate} object, given as a vector, matrix or dataframe (the same type as xv)}

\item{batch_size}{the size of the batches at which to perform the predictions, default is 100.}

\item{...}{additional arguments (although not quite sure what these may be)}
}
\value{
\item{Efx}{BL adjusted expectation for f(x)}
\item{Varfx}{BL adjusted variance for f(x)}
}
\description{
predict the value of a model at some new input locations, along with a measure of uncertainty, using a PPBLE of class \code{emulate}.
}
\examples{
f <- function(x){ x[2] * sin( x[1] ) + x[1] * cos( x[2] ) }
x <- matrix( runif( 20, 0.2, 1.2 ), ncol = 2 )
fx <- apply( x, 1, f )
theta <- c( 0.4, 0.6 )
emulator <- emulate( x = x, fx = fx, CF = GaussianCF,
            CF_para = list( theta = theta, delta = 0.0001 )  )
y <- matrix( runif( 16, 0.2, 1.2 ), ncol = 2 )
predict( emulator, y )
}
